pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cmake>=4.2.0",
    "ndi-python==5.1.1.1",
    "onnx>=1.19.1",
    "onnxruntime-gpu>=1.23.2",
    "opencv-python>=4.12.0.88",
    "python-osc>=1.9.3",
    "ultralytics>=8.3.231",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "psutil>=7.1.3",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch" }]
torchvision = [{ index = "pytorch" }]
